g-drive download helper utility and path append/prepend functions
updating the g-drive download shell utility interface

+++ utils/bashfuncs.sh
@@ -7,12 +7,24 @@
 # drive and then untar them. This will mainly be used to download features
 # and datasets.
 
+################################################################################
+############################# GDrive Downloader ################################
+################################################################################
+# The following is a helper function to download tarballs from the google
+# drive and then untar them. This will mainly be used to download features
+# and datasets.
+
 function gdluntar {
-  # usage: gdluntar FILEID FILENAME FILEURL PTHMD5FILE
+  # usage: gdluntar FILEID FILENAME FILEURL PTHMD5FILE RMAFTERDL ISSMALL
   MYFILEID=$1
   MYTARFILE=$2
   MYURL=$3
   PTHMD5FILE=$4
+  RMAFTERDL=$5
+  ISSMALL=$6
+  
+  [[ "aa"$RMAFTERDL == "aa" ]] && echo "args incomplete" && return 1
+  [[ "aa"$ISSMALL == "aa" ]] && ISSMALL=0
 
   if md5sum -c ${PTHMD5FILE} 2> /dev/null; then
     :
@@ -21,8 +33,11 @@ function gdluntar {
     echo "Attempting a fresh download..."
     echo "Downloading: $MYURL -> $(readlink -m $MYTARFILE)"
     sleep 0.5
-
-    wget --load-cookies /tmp/cookies.txt "https://docs.google.com/uc?export=download&confirm=$(wget --quiet --save-cookies /tmp/cookies.txt --keep-session-cookies --no-check-certificate 'https://docs.google.com/uc?export=download&id='${MYFILEID} -O- | sed -rn 's/.*confirm=([0-9A-Za-z_]+).*/\1\n/p')&id=${MYFILEID}" -O ${MYTARFILE} && rm -rf /tmp/cookies.txt
+    if [[ $ISSMALL == "1" ]]; then
+      wget --no-check-certificate 'https://docs.google.com/uc?export=download&id='"${MYFILEID}" -O ${MYTARFILE} 
+    else
+      wget --load-cookies /tmp/cookies.txt "https://docs.google.com/uc?export=download&confirm=$(wget --quiet --save-cookies /tmp/cookies.txt --keep-session-cookies --no-check-certificate 'https://docs.google.com/uc?export=download&id='${MYFILEID} -O- | sed -rn 's/.*confirm=([0-9A-Za-z_]+).*/\1\n/p')&id=${MYFILEID}" -O ${MYTARFILE} && rm -rf /tmp/cookies.txt
+    fi
 
     if [[ ! -f $MYTARFILE ]]; then
       echo "Downloading Error: could not download the file $MYTARFILE."
@@ -39,11 +54,12 @@ function gdluntar {
       echo "    4. [Optional]: check the md5sums of the extracted files:"
       echo "       md5sum $(readlink -m $PTHMD5FILE)"
       echo "*******************************************************************"
-    elif [[ $MYTARFILE == *.tar ]]; then
+    elif [[ $MYTARFILE == *.tar || $MYTARFILE == *.tar.gz ]]; then
       echo "----------";
       echo "Untarring ${MYTARFILE}:"
-      tar -xvf $MYTARFILE
-      rm $MYTARFILE
+      [[ $MYTARFILE == *.tar ]] && tar -xvf $MYTARFILE
+      [[ $MYTARFILE == *.tar.gz ]] && tar -xzvf $MYTARFILE
+      [[ $RMAFTERDL == 1 ]] && rm $MYTARFILE
       echo "----------";
       if md5sum -c ${PTHMD5FILE}; then
         echo "";
